requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.5
django-crispy-forms==1.14.0
gunicorn==20.1.0
sqlparse==0.4.3
tzdata==2022.7